refactor(dsa): extract KPool-specific backend helpers
